// src/car-point/dto/car-point-request.dto.ts
export class CarPointRequestDto {
  brand: string;
  model: string;
  year: number;
  bodyType: string;
  color: string;
  engine: string;
  mileage: number;
  transmission: string;
  drive: string;
  isNew: boolean;
  numberOfSeats: number;
  condition: string;
  market: string;
  city: string;
  price: number;
}